MVVM split
