+++ src/main/java/org/jabref/gui/mergeentries/MultiMergeEntries.fxml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<DialogPane
+        xmlns="http://javafx.com/javafx"
+        xmlns:fx="http://javafx.com/fxml"
+        fx:controller="org.jabref.gui.mergeentries.MultiMergeEntriesView"
+        prefHeight="500.0"
+        prefWidth="800.0">
+
+    <content>
+        <SplitPane fx:id="splitPane" dividerPositions="0.1, 0.6, 0.3" orientation="HORIZONTAL">
+            <items>
+                <VBox fx:id="labelColumn">
+                    <Label fx:id="supplierLabel" text="Supplier" prefHeight="${topScrollPane.height}" VBox.vgrow="NEVER">
+                        <minHeight><Control fx:constant="USE_PREF_SIZE" /></minHeight>
+                        <maxHeight><Control fx:constant="USE_PREF_SIZE" /></maxHeight>
+                    </Label>
+                    <ScrollPane fx:id="leftScrollPane" hbarPolicy="NEVER" vbarPolicy="NEVER" VBox.vgrow="ALWAYS">
+                        <VBox fx:id="fieldHeader"/>
+                    </ScrollPane>
+                </VBox>
+                <VBox fx:id="optionsColumn">
+                    <ScrollPane fx:id="topScrollPane" hbarPolicy="NEVER" vbarPolicy="NEVER" VBox.vgrow="NEVER">
+                        <HBox fx:id="supplierHeader" />
+                    </ScrollPane>
+                    <ScrollPane fx:id="centerScrollPane" vbarPolicy="NEVER" VBox.vgrow="ALWAYS">
+                        <GridPane fx:id="optionsGrid" />
+                    </ScrollPane>
+                </VBox>
+                <VBox fx:id="entryColumn">
+                    <Label fx:id="entryLabel" text="Entry" prefHeight="${topScrollPane.height}" VBox.vgrow="NEVER">
+                        <minHeight><Control fx:constant="USE_PREF_SIZE" /></minHeight>
+                        <maxHeight><Control fx:constant="USE_PREF_SIZE" /></maxHeight>
+                    </Label>
+                    <ScrollPane fx:id="rightScrollPane" hbarPolicy="NEVER" VBox.vgrow="ALWAYS">
+                        <VBox fx:id="fieldEditor"/>
+                    </ScrollPane>
+                </VBox>
+            </items>
+        </SplitPane>
+    </content>
+
+</DialogPane>
